+++ andes/models/renewable/regcv1.py
@@ -23,7 +23,9 @@ class REGCV1Data(ModelData):
                             mandatory=True,
                             )
         self.gen = IdxParam(info="static generator index",
+                            model='StaticGen',
                             mandatory=True,
+                            replaces=True,
                             )
         self.coi2 = IdxParam(model='COI2',
                              info="center of inertia 2 index",
@@ -271,12 +273,6 @@ class REGCV1ModelBase(Model):
                         v_str='Iq0',
                         diag_eps=True,
                         )
-
-    def v_numeric(self, **kwargs):
-        """
-        Disable the corresponding `StaticGen`s.
-        """
-        self.system.groups['StaticGen'].set(src='u', idx=self.gen.v, attr='v', value=0)
 
 
 class VSGOuterPIModel: