+++ andes/models/synchronous/genbase.py
@@ -17,6 +17,7 @@ class GENBaseData(ModelData):
         self.gen = IdxParam(info="static generator index",
                             model='StaticGen',
                             mandatory=True,
+                            replaces=True,
                             )
         self.coi = IdxParam(model='COI',
                             info="center of inertia index",
@@ -250,14 +251,6 @@ class GENBase(Model):
                         info='reactive power injection',
                         e_str='u * (vq * Id - vd * Iq) - Qe',
                         v_str='u * (vq0 * Id0 - vd0 * Iq0)')
-
-    def v_numeric(self, **kwargs):
-        """
-        Disable static generators with a linked synchronous machine.
-        """
-
-        mask_idx = [self.gen.v[i] for i in range(self.n) if self.u.v[i] == 1]
-        self.system.groups['StaticGen'].set(src='u', idx=mask_idx, attr='v', value=0)
 
 
 class Flux0: